+++ src/main/java/org/scion/jpan/internal/LocalTopology.java
@@ -21,6 +21,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import org.scion.jpan.Constants;
 import org.scion.jpan.ScionRuntimeException;
 import org.scion.jpan.ScionUtil;
 
@@ -46,6 +47,13 @@ public class LocalTopology {
 
     public static DispatcherPortRange createAll() {
       return new DispatcherPortRange(-1, -1);
+    }
+
+    public static DispatcherPortRange createEmpty() {
+      // For the empty range we allow only port 30041.
+      // If the port is used by the SHIM, than we cannot connect. However, if there is no SHIM,
+      // then we can safely use 30041.
+      return new DispatcherPortRange(Constants.SCMP_PORT, Constants.SCMP_PORT);
     }
 
     public boolean hasPortRange() {
@@ -151,7 +159,9 @@ public class LocalTopology {
         }
       }
       JsonElement underlay = o.get("underlay");
-      if (underlay != null) {
+      if (underlay == null) {
+        portRange = DispatcherPortRange.createEmpty();
+      } else {
         JsonObject u = underlay.getAsJsonObject();
         portRange = parsePortRange(u.get("dispatched_ports").getAsString());
       }
@@ -166,9 +176,7 @@ public class LocalTopology {
 
   private static DispatcherPortRange parsePortRange(String v) {
     if ("-".equals(v)) {
-      // EMPTY range means the only port we can possibly use is 30041, if no SHIM is listening
-      // there.
-      return new DispatcherPortRange(30041, 30041);
+      return DispatcherPortRange.createEmpty();
     } else if ("all".equalsIgnoreCase(v)) {
       return DispatcherPortRange.createAll();
     } else {
@@ -178,7 +186,7 @@ public class LocalTopology {
       }
       int portMin = Integer.parseInt(sa[0]);
       int portMax = Integer.parseInt(sa[1]);
-      if (portMin < 1 || portMin > 65535 || portMax < 1 || portMax > 65535 || portMin > portMax) {
+      if (portMin < 1 || portMax < 1 || portMax > 65535 || portMin > portMax) {
         throw new ScionRuntimeException("Illegal port values in topo file dispatched_ports: " + v);
       }
       return new DispatcherPortRange(portMin, portMax);